insomnia: Attempt to resolve the tmux/PATH fuckery
The problem is that tmux clears the environment and spawns a login
shell. The shell set's PATH and thus removes the PATH element containing
our insomnia utility scripts.

The solution implemented in this commit is a script which adds the
directory containing our utility scripts to PATH and afterwards execute
the given command. This means that all utility scripts have to be
started through this shell script, which isn't optimal, but sufficient
for now.