insomnia-output: create outfp if it doesn't exist
Previously a directory always contained an out file because the
directory was only created if a conversation has been started. However,
hii recently gained support for the IRCv3.2 MONITOR command.

This means that a directory might be created for a user with whom a
conversation hasn't been started yet. To handle this special case
properly the file is created by insomnia-output and hii appends content
to it later.

touch(1) is only called if the file doesn't exist to prevent unnecessary
modifications of the mtime. Also change the test from -f to -e.