insomnia-color: rewrite in C
While working on an additional insomnia-pager binary for viewing log
files I noticed that insomnia-color is incredibly slow. The new C
version is noticeably faster:

    $ wc -l < large-out-file
    172463
    $ time sh -c './insomnia-color-awk < large-out-file >/dev/null'
    0m07.11s real     0m07.06s user     0m00.04s system
    $ time sh -c './insomnia-color-c < large-out-file >/dev/null'
    0m00.78s real     0m00.73s user     0m00.05s system

For this reason, the new C version is more suitable for piping the
output of insomnia-color to a pager. Additionally, the code is almost
the same length and the C code even seems a bit cleaner as it doesn't
require an ASCII lookup table.