insomnia: update comment on session name
It would be possible to escape control characters in session names.
However, at this point in time my entire insomnia workflow is based
around the fact that there can only be a single session. Besides, this
reduces the likelihood of pilling up old sessions which stay around
unnoticed and consume CPU resources.