insomnia-color: use more restrictive regex for nick names
With the previously used AWK script this didn't matter much as AWK did
not use a subexpression for extracting the nick name. However, the C
software does and therefore needs to be careful with nested parentheses.
For example:

	1582492312 (test) foo (bar) baz

To workaround this issue do not allow closing parentheses as nick name
characters. This is in accordance with the nickname ABNF from RFC 2812.