insomnia: attempt to resolve some tmux related race conditions
When a tmux command returns it does not seem to guarantee that its
operation was already performed successfully. For instance, after
new-session returns the session may not exist yet, thereby causing
set-hook to run into an error. These kinds of concurrency issues do not
seem to arise when relying on tmux to determine the current session,
i.e. instead of specifying the session explicitly. For this reason, this
commit removes explicit tmux -t options when they are not needed.
Besides, the attach is now also executed as part of the commands
creating the session, again without an explicit -t option. Similarly,
the entire tmux command is now executed using exec(1).