insomnia-sort: Cancel pending alarm after returning from input loop
Technically, there is a race condition if the input stream is finite
(i.e. if inloop() can return) and if the SIGALRM is received exactly
when we return from inloop() execute sortprint() from main() in which
case we would jump to the signal handler and execute it there as well.

Canceling the pending alarm should prevent this race condition.
Alternatively, we could also block the SIGALRM signal after returning
from inloop() and before executing sortprint() in main().